Authenticate via s2s key or admin key or ouath2 token
