+++ papiea-engine/src/auth/oauth2.ts
@@ -1,5 +1,5 @@
-import { NextFunction, Response, Router } from "express";
-import { asyncHandler, UnauthorizedError, UserAuthInfo, UserAuthInfoRequest } from "./authn";
+import { Router } from "express";
+import { asyncHandler, UserAuthInfo } from "./authn";
 import { Signature } from "./crypto";
 import { Provider_DB } from "../databases/provider_db_interface";
 import { extract_property } from "./user_data_evaluator";
@@ -45,18 +45,7 @@ function getUserInfoFromToken(token: any, provider: Provider): UserAuthInfo {
     return userInfo;
 }
 
-function getToken(req: any): string | null {
-    if (req.headers.authorization && req.headers.authorization.split(' ')[0] === 'Bearer') {
-        return req.headers.authorization.split(' ')[1];
-    } else if (req.query && req.query.token) {
-        return req.query.token;
-    } else if (req.cookies && req.cookies.token) {
-        return req.cookies.token;
-    }
-    return null;
-}
-
-export function createOAuth2Router(redirect_uri: string, signature: Signature, providerDb: Provider_DB) {
+export function createOAuth2Router(redirect_uri: string, signature: Signature, providerDb: Provider_DB): Router {
     const router = Router();
 
     router.use('/provider/:prefix/:version/auth/login', asyncHandler(async (req, res) => {
@@ -103,31 +92,6 @@ export function createOAuth2Router(redirect_uri: string, signature: Signature, p
             console.error('Access Token Error', error.message);
             return res.status(500).json('Authentication failed');
         }
-    }));
-
-    async function injectUserInfo(req: UserAuthInfoRequest, res: Response, next: NextFunction): Promise<void> {
-        const token = getToken(req);
-        if (token === null) {
-            return next();
-        }
-        let userInfo: UserAuthInfo;
-        try {
-            userInfo = await signature.verify(token);
-        } catch (e) {
-            throw new UnauthorizedError();
-        }
-        if (userInfo.provider_prefix !== req.params.prefix) {
-            throw new UnauthorizedError();
-        }
-        req.user = userInfo;
-        next();
-    }
-
-    router.use('/services/:prefix', asyncHandler(injectUserInfo));
-    router.use('/provider/:prefix', asyncHandler(injectUserInfo));
-
-    router.use('/provider/:prefix/:version/auth/user_info', asyncHandler(async (req, res) => {
-        res.json(req.user);
     }));
 
     return router;
